+++ src/utils/exception.h
@@ -18,10 +18,29 @@ namespace vroom {
 
 class Exception : public std::exception {
 public:
-  const ERROR error;
   const std::string message;
+  unsigned int error_code;
 
-  Exception(ERROR error, const std::string& message);
+  Exception(const std::string& message, unsigned int error_code);
+
+  const char* what() const noexcept override {
+    return message.c_str();
+  };
+};
+
+class InternalException : public Exception {
+public:
+  InternalException(const std::string& message);
+};
+
+class InputException : public Exception {
+public:
+  InputException(const std::string& message);
+};
+
+class RoutingException : public Exception {
+public:
+  RoutingException(const std::string& message);
 };
 
 } // namespace vroom